requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.1.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Faker==13.3.4
idna==3.3
kombu==5.2.4
lorem==0.1.1
names==0.3.0
packaging==21.3
Pillow==9.0.0
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2021.3
redis==4.2.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.3.2
wrapt==1.14.0

aiohttp~=3.8.1
django-debug-toolbar~=3.2.4